# src/backend/routes/servicos.js
const express = require("express");
const { Servico, Veiculo, Cliente, ItemServico, Produto } = require("../models");
const { isIntegerGreaterThanZero } = require("../utils/utils");

const router = express.Router();

const servicoInclude = [
  {
    model: Veiculo,
    as: "veiculo",
    include: [
      {
        model: Cliente,
        as: "cliente"
      }
    ]
  },
  {
    model: ItemServico,
    as: "itens",
    include: [
      {
        model: Produto,
        as: "produto"
      }
    ]
  }
];

async function buscarServicoCompleto(id) {
  return Servico.findByPk(id, {
    include: servicoInclude
  });
}

router.get("/", async (req, res) => {
  try {
    const servicos = await Servico.findAll({
      include: servicoInclude,
      order: [["id", "DESC"]]
    });

    return res.json(servicos);
  } catch (error) {
    return res.status(500).json({
      message: "Erro interno ao buscar serviços"
    });
  }
});

router.get("/:id", async (req, res) => {
  try {
    const id = Number(req.params.id);

    const idValido = isIntegerGreaterThanZero(id);

    if (!idValido) {
      return res.status(400).json({
        message: "ID deve ser um número inteiro maior que zero"
      });
    }

    const servico = await buscarServicoCompleto(id);

    if (!servico) {
      return res.status(404).json({
        message: "Serviço não encontrado"
      });
    }

    return res.json(servico);
  } catch (error) {
    return res.status(500).json({
      message: "Erro interno ao buscar serviço"
    });
  }
});

router.post("/", async (req, res) => {
  try {
    const {
      descricao,
      status,
      dataInicio,
      dataFim,
      valorTotal: valorTotalPayload,
      idVeiculo: idVeiculoPayload
    } = req.body;

    const idVeiculo = Number(idVeiculoPayload);
    const valorTotal = valorTotalPayload === undefined ? 0 : Number(valorTotalPayload);

    if (!descricao || !status || !dataInicio || !idVeiculoPayload) {
      return res.status(400).json({
        message: "descricao, status, dataInicio e idVeiculo são obrigatórios"
      });
    }

    if (!isIntegerGreaterThanZero(idVeiculo)) {
      return res.status(400).json({
        message: "idVeiculo deve ser um número inteiro maior que zero"
      });
    }

    if (Number.isNaN(valorTotal) || valorTotal < 0) {
      return res.status(400).json({
        message: "valorTotal deve ser um número maior ou igual a zero"
      });
    }

    const veiculo = await Veiculo.findByPk(idVeiculo);

    if (!veiculo) {
      return res.status(404).json({
        message: "Veículo não encontrado"
      });
    }

    const servico = await Servico.create({
      descricao,
      status,
      dataInicio,
      dataFim: dataFim ?? null,
      valorTotal: valorTotal,
      idVeiculo: idVeiculo
    });

    const servicoCompleto = await buscarServicoCompleto(servico.id);

    return res.status(201).json(servicoCompleto);
  } catch (error) {
    if (error.name === "SequelizeValidationError") {
      return res.status(400).json({
        message: "Erro de validação",
        errors: error.errors.map((err) => ({
          field: err.path,
          message: err.message
        }))
      });
    }

    if (error.name === "SequelizeForeignKeyConstraintError") {
      return res.status(400).json({
        message: "Veículo informado não existe"
      });
    }

    return res.status(500).json({
      message: "Erro interno ao cadastrar serviço"
    });
  }
});

router.put("/:id", async (req, res) => {
  try {
    const id = Number(req.params.id);

    const idValido = isIntegerGreaterThanZero(id);

    if (!idValido) {
      return res.status(400).json({
        message: "ID deve ser um número inteiro maior que zero"
      });
    }

    const servico = await Servico.findByPk(id);

    if (!servico) {
      return res.status(404).json({
        message: "Serviço não encontrado"
      });
    }

    const {
      descricao,
      status,
      dataInicio,
      dataFim,
      valorTotal: valorTotalPayload,
      idVeiculo: idVeiculoPayload
    } = req.body;

    let idVeiculo = servico.idVeiculo;
    let valorTotal = servico.valorTotal;

    if (idVeiculoPayload !== undefined) {
      idVeiculo = Number(idVeiculoPayload);

      if (!isIntegerGreaterThanZero(idVeiculo)) {
        return res.status(400).json({
          message: "idVeiculo deve ser um número inteiro maior que zero"
        });
      }

      const veiculo = await Veiculo.findByPk(idVeiculo);

      if (!veiculo) {
        return res.status(404).json({
          message: "Veículo não encontrado"
        });
      }
    }

    if (valorTotalPayload !== undefined) {
      valorTotal = Number(valorTotalPayload);

      if (Number.isNaN(valorTotal) || valorTotal < 0) {
        return res.status(400).json({
          message: "valorTotal deve ser um número maior ou igual a zero"
        });
      }
    }

    const servicoAtualizado = await servico.update({
      descricao: descricao ?? servico.descricao,
      status: status ?? servico.status,
      dataInicio: dataInicio ?? servico.dataInicio,
      dataFim: dataFim ?? servico.dataFim,
      valorTotal: valorTotal,
      idVeiculo: idVeiculo
    });

    const servicoCompleto = await buscarServicoCompleto(servicoAtualizado.id);

    return res.json(servicoCompleto);
  } catch (error) {
    if (error.name === "SequelizeValidationError") {
      return res.status(400).json({
        message: "Erro de validação",
        errors: error.errors.map((err) => ({
          field: err.path,
          message: err.message
        }))
      });
    }

    if (error.name === "SequelizeForeignKeyConstraintError") {
      return res.status(400).json({
        message: "Veículo informado não existe"
      });
    }

    return res.status(500).json({
      message: "Erro interno ao atualizar serviço"
    });
  }
});

router.delete("/:id", async (req, res) => {
  try {
    const id = Number(req.params.id);

    const idValido = isIntegerGreaterThanZero(id);

    if (!idValido) {
      return res.status(400).json({
        message: "ID deve ser um número inteiro maior que zero"
      });
    }

    const servico = await Servico.findByPk(id);

    if (!servico) {
      return res.status(404).json({
        message: "Serviço não encontrado"
      });
    }

    const possuiItens = await ItemServico.findOne({
      where: {
        idServico: id
      }
    });

    if (possuiItens) {
      return res.status(409).json({
        message: "Serviço possui itens vinculados"
      });
    }

    await servico.destroy();

    return res.status(200).json({
      message: "Serviço deletado com sucesso"
    });
  } catch (error) {
    return res.status(500).json({
      message: "Erro interno ao deletar serviço"
    });
  }
});

module.exports = router;