# src/backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AutoPro API",
    "description": "Documentacao das rotas da API AutoPro baseada na modelagem Cliente, Veiculo, Servico, Item de Servico e Produto.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Servidor local"
    }
  ],
...
  "paths": {
...
    "/servicos": {
      "get": {
        "summary": "Lista todos os servicos",
        "tags": [
          "Servicos"
        ],
        "responses": {
          "200": {
            "description": "Lista de servicos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Servico"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Cria um servico",
        "tags": [
          "Servicos"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServicoInput"
              },
              "example": {
                "descricao": "Troca de oleo",
                "status": "em andamento",
                "dataInicio": "2026-05-01",
                "dataFim": null,
                "valorTotal": 250,
                "idVeiculo": 1
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Servico criado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Servico"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/servicos/{id}": {
      "get": {
        "summary": "Busca um servico por ID",
        "tags": [
          "Servicos"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Id"
          }
        ],
        "responses": {
          "200": {
            "description": "Servico encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Servico"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "put": {
        "summary": "Atualiza um servico",
        "tags": [
          "Servicos"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Id"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServicoInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Servico atualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Servico"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "summary": "Remove um servico",
        "tags": [
          "Servicos"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Id"
          }
        ],
        "responses": {
          "204": {
            "description": "Servico removido"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
...
    "/produtos": {
      "get": {
        "summary": "Lista todos os produtos",
        "tags": [
          "Produtos"
        ],
        "responses": {
          "200": {
            "description": "Lista de produtos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Produto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Cria um produto",
        "tags": [
          "Produtos"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProdutoInput"
              },
              "example": {
                "nome": "Oleo 5W30",
                "quantidade": 12,
                "precoUnitario": 65.5
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Produto criado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Produto"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/produtos/{id}": {
      "get": {
        "summary": "Busca um produto por ID",
        "tags": [
          "Produtos"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Id"
          }
        ],
        "responses": {
          "200": {
            "description": "Produto encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Produto"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "put": {
        "summary": "Atualiza um produto",
        "tags": [
          "Produtos"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Id"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProdutoInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Produto atualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Produto"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "summary": "Remove um produto",
        "tags": [
          "Produtos"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Id"
          }
        ],
        "responses": {
          "204": {
            "description": "Produto removido"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "Id": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "ID do registro",
        "schema": {
          "type": "integer",
          "example": 1
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Dados invalidos ou obrigatorios ausentes",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiMessage"
            }
          }
        }
      },
      "NotFound": {
        "description": "Registro nao encontrado",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiMessage"
            }
          }
        }
      },
      "Conflict": {
        "description": "Operacao bloqueada por relacionamento ou regra de negocio",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiMessage"
            }
          }
        }
      },
...
    },
    "schemas": {
      "ApiMessage": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Mensagem da API"
          }
        }
      },
...
      "Servico": {
        "allOf": [
          {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "example": 1
              }
            }
          },
          {
            "$ref": "#/components/schemas/ServicoInput"
          }
        ]
      },
      "ServicoInput": {
        "type": "object",
        "required": [
          "descricao",
          "status",
          "dataInicio",
          "valorTotal",
          "idVeiculo"
        ],
        "properties": {
          "descricao": {
            "type": "string",
            "example": "Troca de oleo"
          },
          "status": {
            "type": "string",
            "example": "em andamento"
          },
          "dataInicio": {
            "type": "string",
            "format": "date",
            "example": "2026-05-01"
          },
          "dataFim": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "example": "2026-05-02"
          },
          "valorTotal": {
            "type": "number",
            "format": "float",
            "example": 250
          },
          "idVeiculo": {
            "type": "integer",
            "example": 1
          }
        }
      },
...
      "Produto": {
        "allOf": [
          {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "example": 1
              }
            }
          },
          {
            "$ref": "#/components/schemas/ProdutoInput"
          }
        ]
      },
      "ProdutoInput": {
        "type": "object",
        "required": [
          "nome",
          "quantidade",
          "precoUnitario"
        ],
        "properties": {
          "nome": {
            "type": "string",
            "example": "Oleo 5W30"
          },
          "quantidade": {
            "type": "number",
            "format": "float",
            "example": 12
          },
          "precoUnitario": {
            "type": "number",
            "format": "float",
            "example": 65.5
          }
        }
      },
...
    }
  }
}